Move experimental "indent" rule to standard one.
