+++ ktlint-ruleset-standard/src/main/kotlin/com/pinterest/ktlint/ruleset/standard/StandardRuleSetProvider.kt
@@ -13,6 +13,7 @@ class StandardRuleSetProvider : RuleSetProvider {
         FilenameRule(),
         FinalNewlineRule(),
         ImportOrderingRule(),
+        IndentationRule(),
         MaxLineLengthRule(),
         ModifierOrderRule(),
         NoBlankLineBeforeRbraceRule(),
