+++ ktlint-ruleset-standard/src/test/kotlin/com/pinterest/ktlint/ruleset/standard/IndentationRuleTest.kt
@@ -1,4 +1,4 @@
-package com.pinterest.ktlint.ruleset.experimental
+package com.pinterest.ktlint.ruleset.standard
 
 import com.pinterest.ktlint.core.LintError
 import com.pinterest.ktlint.test.diffFileFormat
@@ -8,7 +8,7 @@ import com.pinterest.ktlint.test.lint
 import org.assertj.core.api.Assertions.assertThat
 import org.junit.Test
 
-class IndentationRuleTest {
+internal class IndentationRuleTest {
 
     @Test
     fun testLint() {
