+++ ktlint-ruleset-standard/src/test/resources/spec/indent/format-argument-list.kt.spec
