+++ ktlint-ruleset-standard/src/test/resources/spec/indent/format-arrow-expected.kt.spec
