+++ ktlint-ruleset-standard/src/test/resources/spec/indent/format-binary-expression.kt.spec
