+++ ktlint-ruleset-standard/src/test/resources/spec/indent/format-dot-qualified-expression.kt.spec
