+++ ktlint-ruleset-standard/src/test/resources/spec/indent/format-eq-expected.kt.spec
