+++ ktlint-ruleset-standard/src/test/resources/spec/indent/format-expected.kt.spec
