+++ ktlint-ruleset-standard/src/test/resources/spec/indent/format-kdoc-expected.kt.spec
