+++ ktlint-ruleset-standard/src/test/resources/spec/indent/format-multiline-string.kt.spec
