+++ ktlint-ruleset-standard/src/test/resources/spec/indent/format-parameter-list-expected.kt.spec
