+++ ktlint-ruleset-standard/src/test/resources/spec/indent/format-property-accessor.kt.spec
