+++ ktlint-ruleset-standard/src/test/resources/spec/indent/format-raw-string-trim-indent.kt.spec
