+++ ktlint-ruleset-standard/src/test/resources/spec/indent/format-supertype-expected.kt.spec
