+++ ktlint-ruleset-standard/src/test/resources/spec/indent/format.kt.spec
