+++ ktlint-ruleset-standard/src/test/resources/spec/indent/lint-argument-list.kt.spec
