+++ ktlint-ruleset-standard/src/test/resources/spec/indent/lint-binary-expression.kt.spec
