+++ ktlint-ruleset-standard/src/test/resources/spec/indent/lint-comment.kt.spec
