+++ ktlint-ruleset-standard/src/test/resources/spec/indent/lint-condition.kt.spec
