+++ ktlint-ruleset-standard/src/test/resources/spec/indent/lint-dot-qualified-expression.kt.spec
