+++ ktlint-ruleset-standard/src/test/resources/spec/indent/lint-property-accessor.kt.spec
