+++ ktlint-ruleset-standard/src/test/resources/spec/indent/lint-string.kt.spec
