+++ ktlint-ruleset-standard/src/test/resources/spec/indent/lint-supertype.kt.spec
