+++ ktlint-ruleset-standard/src/test/resources/spec/indent/lint-when-expression.kt.spec
