+++ ktlint-ruleset-standard/src/test/resources/spec/indent/lint.kt.spec
